Fix the name of the install plan to approve
If the resource name is not specified, the policy engine refuses to update the resource and turns the enforme remediation policy to inform.

Removed the configuration policy to check that the install plan existed.  The install plan approval depends on the subscription now

The name of the configuration policies cannot contain upper case letters.

This policy is still not working properly, when a new install plan appears, it is approved.

Signed-off-by: GTPE Student <lab-user@bastion.ocp4-cluster-fqgtk.svc.cluster.local>