Added logging policies
Added a policy to install the logging operator

Added a policy to create the Loki Stack custom resource

To create the secret to access the S3 bucket (OBC) the go template function "cat" is used, but this function does not allow the use of other functions inside, so a variable definition needs to be used instead.  The replace function is used to remove extra whitespace created by the cat function, and the resutl is converted to base64 so it can be used in the secret definition directly.

Signed-off-by: GTPE Student <lab-user@bastion.t2dcq.internal>