Added data retention policy
A Job and config map are created to runn a script that uploads the lifecycle configuration to the S3 bucket.
This configuration defines a retention policy of 1 day.
The scripts to get and set the lifecycle and some additional information can be found here: https://github.com/tale-toul/S3lifecycle

Signed-off-by: GTPE Student <lab-user@bastion.q74d9.internal>